r Restructures the node package and improves its documentation on how to use its API
Introducing Profiles: Replace compilation targets with profiles #55
