+++ packages/node/package.json
@@ -4,7 +4,6 @@
     "description": "Node library for running websmith with CLI and webpack in Node.js",
     "author": "Quatico Solutions AG",
     "license": "MIT",
-    "private": true,
     "main": "src/index.ts",
     "publishConfig": {
         "main": "lib/index.js",
@@ -15,6 +14,14 @@
         "README.md",
         "lib"
     ],
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/quatico-solutions/websmith.git"
+    },
+    "bugs": {
+        "url": "https://github.com/quatico-solutions/websmith/issues"
+    },
+    "homepage": "https://github.com/quatico-solutions/websmith#readme",
     "scripts": {
         "clean": "rimraf lib bin coverage",
         "lint": "eslint \"{src,test}/**/*.ts\" --color",
@@ -68,6 +75,7 @@
         "tildify": "^3.0.0"
     },
     "peerDependencies": {
+        "@quatico/websmith-core": "workspace:*",
         "ts-loader": "9.x",
         "typescript": "5.x",
         "webpack": "5.x"
