+++ packages/node/src/websmith/Compiler.ts
