+++ packages/node/src/websmith/compile.spec.ts
