+++ packages/node/src/websmith/compile.ts
