+++ packages/node/src/websmith/console-output.spec.ts
