+++ packages/node/src/websmith/console-output.ts
