+++ packages/node/src/websmith/find-tsc.ts
